+++ client/context/query.go
@@ -3,7 +3,6 @@ package context
 import (
 	"fmt"
 	"strings"
-	"time"
 
 	"github.com/pkg/errors"
 
@@ -13,7 +12,6 @@ import (
 	tmliteErr "github.com/tendermint/tendermint/lite/errors"
 	tmliteProxy "github.com/tendermint/tendermint/lite/proxy"
 	rpcclient "github.com/tendermint/tendermint/rpc/client"
-	ctypes "github.com/tendermint/tendermint/rpc/core/types"
 	tmtypes "github.com/tendermint/tendermint/types"
 
 	"github.com/cosmos/cosmos-sdk/store/rootmulti"
@@ -28,55 +26,6 @@ func (ctx CLIContext) GetNode() (rpcclient.Client, error) {
 	}
 
 	return ctx.Client, nil
-}
-
-// WaitForNBlocks blocks until the node defined on the context has advanced N blocks
-func (ctx CLIContext) WaitForNBlocks(n int64) {
-	node, err := ctx.GetNode()
-	if err != nil {
-		panic(err)
-	}
-
-	resBlock, err := node.Block(nil)
-	var height int64
-	if err != nil || resBlock.Block == nil {
-		// wait for the first block to exist
-		ctx.waitForHeight(1)
-		height = 1 + n
-	} else {
-		height = resBlock.Block.Height + n
-	}
-	ctx.waitForHeight(height)
-}
-
-func (ctx CLIContext) waitForHeight(height int64) {
-	node, err := ctx.GetNode()
-	if err != nil {
-		panic(err)
-	}
-
-	for {
-		// get url, try a few times
-		var resBlock *ctypes.ResultBlock
-		var err error
-	INNER:
-		for i := 0; i < 5; i++ {
-			resBlock, err = node.Block(nil)
-			if err == nil {
-				break INNER
-			}
-			time.Sleep(time.Millisecond * 200)
-		}
-		if err != nil {
-			panic(err)
-		}
-
-		if resBlock.Block != nil && resBlock.Block.Height >= height {
-			return
-		}
-
-		time.Sleep(time.Millisecond * 100)
-	}
 }
 
 // Query performs a query to a Tendermint node with the provided path.
@@ -178,13 +127,13 @@ func (ctx CLIContext) queryABCI(req abci.RequestQuery) (abci.ResponseQuery, erro
 // or an error if the query fails. In addition, it will verify the returned
 // proof if TrustNode is disabled. If proof verification fails or the query
 // height is invalid, an error will be returned.
-func (ctx CLIContext) query(path string, key tmbytes.HexBytes) (res []byte, height int64, err error) {
+func (ctx CLIContext) query(path string, key tmbytes.HexBytes) ([]byte, int64, error) {
 	resp, err := ctx.queryABCI(abci.RequestQuery{
 		Path: path,
 		Data: key,
 	})
 	if err != nil {
-		return
+		return nil, 0, err
 	}
 
 	return resp.Value, resp.Height, nil