+++ client/keys/add.go
@@ -39,42 +39,8 @@ const (
 	DefaultKeyPass = "12345678"
 )
 
-func addKeyScriptingCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "test [name] [mnemonic] [password]",
-		Short: "Add a recovered mnemonic to the keystore encrypt it, and save to disk, for testing",
-		Long: `Derive a new private key from an existing mnemonic file and encrypt to disk.
-
-NOTE: This is insecure and only meant to be used during testing!!!! HERE BE DRAGONS!		
-`,
-		Args: cobra.ExactArgs(3),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			kb, err := NewKeyBaseFromDir(viper.GetString(flags.FlagHome))
-			if err != nil {
-				return err
-			}
-
-			_, err = kb.Get(args[0])
-			if err == nil {
-				return errors.New("key already exists, exiting")
-			}
-
-			if !bip39.IsMnemonicValid(args[1]) {
-				return errors.New("invalid mnemonic")
-			}
-
-			info, err := kb.CreateAccount(args[0], args[1], "", args[2], keys.CreateHDPath(0, 0).String(), keys.Secp256k1)
-			if err != nil {
-				return err
-			}
-
-			return printCreate(cmd, info, false, "")
-		},
-	}
-	return cmd
-}
-
-func addKeyCommand() *cobra.Command {
+// AddKeyCommand defines a keys command to add a generated or recovered private key to keybase.
+func AddKeyCommand() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "add <name>",
 		Short: "Add an encrypted private key (either newly generated or recovered), encrypt it, and save to disk",